Revert "rename typing.py to typehints.py"
This reverts commit ff305cfec3970b8b448b6c11dfa7be9e763af915.
