+++ audformat/core/typing.py
